Move correction out of folder
